# fern/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Xcelsior",
    "description": "Distributed GPU orchestration platform for AI/ML workloads. Canadian-first data sovereignty (PIPEDA/Law 25), automated trust scoring, Stripe Connect marketplace billing, and real-time GPU telemetry.\n\n**Authentication**: Bearer token via `Authorization: Bearer <token>` header.\n\n**Regions**: Canada-first with province-level filtering (ON, QC, BC, AB, etc.).\n\n**SDK**: Generate idiomatic Python/TypeScript SDKs with [Fern](https://buildwithfern.com).\n\n**LLM Integration**: See `/llms.txt` for AI agent–optimized documentation.\n\nThis published OpenAPI document contains the public developer surface only. Account settings, internal worker, admin, maintenance, and callback endpoints are intentionally omitted.",
    "contact": {
      "name": "Xcelsior",
      "url": "https://xcelsior.ca/",
      "email": "admin@xcelsior.ca"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.3.0"
  },
  "servers": [
    {
      "url": "https://xcelsior.ca",
      "description": "Production"
    },
    {
      "url": "http://localhost:9500",
      "description": "Development"
    }
  ],
  "paths": {
...
    "/v1/inference": {
      "post": {
        "tags": [
          "Inference v2"
        ],
        "summary": "Api V1 Inference Sync",
        "description": "Synchronous inference endpoint (OpenAI-compatible path).\n\nSubmits an inference request and polls for results up to 30 seconds.\nIf stream=true, returns SSE text/event-stream with token deltas.",
        "operationId": "api_v1_inference_sync_v1_inference_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1InferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/inference/async": {
      "post": {
        "tags": [
          "Inference v2"
        ],
        "summary": "Api V1 Inference Async",
        "description": "Asynchronous inference — returns job_id immediately for polling.",
        "operationId": "api_v1_inference_async_v1_inference_async_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1InferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/inference/{job_id}": {
      "get": {
        "tags": [
          "Inference v2"
        ],
        "summary": "Api V1 Inference Poll",
        "description": "Poll for inference results by job_id.",
        "operationId": "api_v1_inference_poll_v1_inference__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/inference/endpoints": {
      "get": {
        "tags": [
          "Inference v2"
        ],
        "summary": "Api Inference List Endpoints",
        "description": "List inference endpoints for the current user.",
        "operationId": "api_inference_list_endpoints_api_v2_inference_endpoints_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Inference v2"
        ],
        "summary": "Api Inference Create Endpoint",
        "description": "Create a serverless inference endpoint.\n\nIf min_workers >= 1, a worker container is provisioned immediately.\nBilling is real-time from credits (per-second compute + per-token inference).",
        "operationId": "api_inference_create_endpoint_api_v2_inference_endpoints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferenceEndpointCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/inference/endpoints/{endpoint_id}": {
      "get": {
        "tags": [
          "Inference v2"
        ],
        "summary": "Api Inference Get Endpoint",
        "description": "Get inference endpoint details.",
        "operationId": "api_inference_get_endpoint_api_v2_inference_endpoints__endpoint_id__get",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Endpoint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Inference v2"
        ],
        "summary": "Api Inference Delete Endpoint",
        "description": "Delete an inference endpoint.",
        "operationId": "api_inference_delete_endpoint_api_v2_inference_endpoints__endpoint_id__delete",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Endpoint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/inference/endpoints/{endpoint_id}/usage": {
      "get": {
        "tags": [
          "Inference v2"
        ],
        "summary": "Api Inference Endpoint Usage",
        "description": "Get usage stats for an inference endpoint.",
        "operationId": "api_inference_endpoint_usage_api_v2_inference_endpoints__endpoint_id__usage_get",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Endpoint Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "Inference v2"
        ],
        "summary": "Api Openai Chat Completions",
        "description": "OpenAI-compatible chat completions endpoint.\n\nRoutes to a serverless inference endpoint running the requested model.",
        "operationId": "api_openai_chat_completions_v1_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InferenceEndpointCreate": {
        "properties": {
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "gpu_type": {
            "type": "string",
            "title": "Gpu Type",
            "default": ""
          },
          "region": {
            "type": "string",
            "title": "Region",
            "default": "ca-east"
          },
          "docker_image": {
            "type": "string",
            "title": "Docker Image",
            "default": "xcelsior/vllm:latest"
          },
          "min_workers": {
            "type": "integer",
            "title": "Min Workers",
            "default": 0
          },
          "max_workers": {
            "type": "integer",
            "title": "Max Workers",
            "default": 3
          },
          "max_batch_size": {
            "type": "integer",
            "title": "Max Batch Size",
            "default": 8
          },
          "max_concurrent": {
            "type": "integer",
            "title": "Max Concurrent",
            "default": 4
          },
          "scaledown_window_sec": {
            "type": "integer",
            "title": "Scaledown Window Sec",
            "default": 300
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "sync"
          },
          "health_endpoint": {
            "type": "string",
            "title": "Health Endpoint",
            "default": "/health"
          },
          "api_format": {
            "type": "string",
            "title": "Api Format",
            "default": "openai"
          }
        },
        "type": "object",
        "required": [
          "model_name"
        ],
        "title": "InferenceEndpointCreate"
      },
...
      "V1InferenceRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Model name or HuggingFace repo"
          },
          "inputs": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "string"
              }
            ],
            "title": "Inputs",
            "description": "Text input(s) for inference"
          },
          "max_tokens": {
            "type": "integer",
            "maximum": 8192.0,
            "minimum": 1.0,
            "title": "Max Tokens",
            "default": 512
          },
          "temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 1.0
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "description": "Stream response via SSE",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "model",
          "inputs"
        ],
        "title": "V1InferenceRequest",
        "description": "OpenAI-compatible inference request for /v1/inference."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}